requirements.txt
python-docx==0.8.11
reportlab==4.0.7
pypdf==3.17.1
PyPDF2==4.1.1
pdfplumber==0.10.3
Pillow==10.0.1
PyQt5==5.15.9
PyQt5-sip==12.13.0
cryptography==41.0.7
pyinstaller==6.1.0
